Add Collapsible component and use it in Asset and Input details pages
