+++ packages/html-vue/src/components/common/Tabs.vue
@@ -1,17 +1,17 @@
 <template>
 	<div>
 		<!-- Tabs list -->
-		<ul
+		<div
 			role="tablist"
-			class="flex items-center gap-2 px-2 w-full border-b border-gray-300 overflow-x-auto"
+			class="flex items-center gap-2 p-1 rounded-lg border border-gray-200 bg-gray-50 overflow-x-auto overflow-y-hidden"
 		>
-			<li
+			<button
 				v-for="option in props.options"
 				:key="option.value"
 				:disabled="option.disabled"
 				:class="[
-					'flex items-center justify-center gap-2 py-2.5 px-6 -mb-0.25 text-sm rounded-t-lg border text-gray-500 outline-hidden cursor-pointer select-none',
-					model === option.value ? 'border-gray-300 text-gray-900 bg-gray-50' : 'border-transparent',
+					'flex items-center justify-center gap-2 py-2 px-6 text-sm rounded-md border text-gray-500 outline-hidden cursor-pointer select-none focus:border-gray-300 hover:border-gray-300',
+					model === option.value ? 'border-gray-300 text-gray-700 bg-white shadow-xs' : 'border-transparent',
 					'transition-colors duration-150',
 					'disabled:opacity-50 disabled:cursor-not-allowed'
 				]"
@@ -19,8 +19,8 @@
 				@click="() => model = option.value"
 			>
 				{{ option.label }}
-			</li>
-		</ul>
+			</button>
+		</div>
 
 		<!-- Tab content -->
 		 <div
